Add iFixit as a sponsor; reword to "supporters"
Yay!  iFixit has offered to sponsor us!  We're lucky to have so many
awesome people supporting us.

Also, the city is not officially a sponsor, since they are contributing
resources, but not funds that we can use at our discretion.  I hope that
calling them a "supporter" (and clarifying that they're a co-organizer)
is OK.  (If not, we'll change it.)